Mejora de modulos

+++ Proyecto-POA/app/Http/Controllers/DatosbasicosController.php
@@ -8,13 +8,13 @@ use App\Models\Datosbasicos;
 class DatosbasicosController extends Controller
 {
     // Función que muestra el contenido del data table
-    public function index()
+    public function index(Request $request)
     {
-        // if(!$request->ajax()) return redirect('/');
+        if(!$request->ajax()) return redirect('/');
         $datobasicos = Datosbasicos::select('dato_basicos.id', 'dato_basicos.denoproceso', 'dato_basicos.numproceso', 
                             'dato_basicos.mejoraproceso', 'dato_basicos.arearesponsable', 'dato_basicos.objetivoproceso',
                             'dato_basicos.pideatencion', 'dato_basicos.problematica')
-                            ->orderBy('dato_basicos.id', 'desc')->paginate(2);
+                            ->orderBy('dato_basicos.id', 'desc')->paginate(1);
         
         return [
             'pagination' => [
@@ -32,7 +32,7 @@ class DatosbasicosController extends Controller
     // Función que contiene la lógica para almacenar un registro
     public function store(Request $request)
     {
-        // if(!$request->ajax()) return redirect('/');
+        if(!$request->ajax()) return redirect('/');
         $datobasico = new Datosbasicos();
         $datobasico->denoproceso = $request->denoproceso;
         $datobasico->numproceso = $request->numproceso;
@@ -47,7 +47,7 @@ class DatosbasicosController extends Controller
     // Función que contiene la lógica para actualizar un registro
     public function update(Request $request)
     {
-        // if(!$request->ajax()) return redirect('/');
+        if(!$request->ajax()) return redirect('/');
         $datobasico = Datosbasicos::findOrFail($request->id);
         $datobasico->denoproceso = $request->denoproceso;
         $datobasico->numproceso = $request->numproceso;
@@ -62,7 +62,7 @@ class DatosbasicosController extends Controller
     // Función que contiene la lógica para eliminar un registro
     public function destroy(Request $request)
     {
-        // if(!$request->ajax()) return redirect('/');
+        if(!$request->ajax()) return redirect('/');
         $datobasico = Datosbasicos::findOrFail($request->id);
         $datobasico->delete();
     }
